requirements.txt
asgiref==3.5.0
aspose-words==22.4.0
Django==4.0.3
django-environ==0.8.1
django-sendfile==0.3.11
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
sqlparse==0.4.2
tk==0.1.0
tzdata==2022.1
